+++ memory.c
@@ -9,12 +9,12 @@
  * @cols: the three levels of expertise.
  * Return: the location of the 2-D array.
  */
-int **allocate_2D_array(unsigned long long int rows, unsigned int cols)
+int **allocate_2D_array(unsigned long int rows, unsigned int cols)
 {
 	int **arrays2D = NULL;
-	unsigned long long int i, j;
+	unsigned long int i, j;
 
-	arrays2D = malloc(sizeof(unsigned long long int) * rows);
+	arrays2D = malloc(sizeof(unsigned long int) * rows);
 	if (!arrays2D)
 	{
 		printf("Failed to allocate memory\n");
@@ -41,9 +41,9 @@ int **allocate_2D_array(unsigned long long int rows, unsigned int cols)
  * @array2D: the array whose members and itself is to be deallocated.
  * @rowSize: the number of rows of the 2-D array.
  */
-void free_2D_array(int **array2D, unsigned long long int rowSize)
+void free_2D_array(int **array2D, unsigned long int rowSize)
 {
-	unsigned long long int i;
+	unsigned long int i;
 
 	for (i = 0; i < rowSize; i++)
 		free(array2D[i]);
